move case_importer_v1 tests to a dir
